+++ src/tests/perception_module_test/DefaultSensorTest.cpp
@@ -101,7 +101,7 @@ BOOST_FIXTURE_TEST_CASE( population_flows_in_list_are_identified_with_default_se
                                                                                            [ sword::test::MakeModel( "x", 1. )( "y", 1. ) ]
                                                                                            [ sword::test::MakeModel( "x", 2. )( "y", 0. ) ] ) ] );
     ExpectEvent( "report", sword::test::MakeModel( "entity", identifier )
-                                                 ( "code", MIL_Report::eReport_CiviliansEncountered ) );
+                                                 ( "code", static_cast< int >( MIL_Report::eReport_CiviliansEncountered ) ) );
     PostCommand( "perception", core::MakeModel( "identifier", identifier ) );
     ExecuteCommands();
 }
@@ -120,7 +120,7 @@ BOOST_FIXTURE_TEST_CASE( population_concentrations_in_list_are_identified_with_d
                                                                                                       ( "level", 3 ) // identified
                                                                                                       ( "recorded", false ) ] );
     ExpectEvent( "report", sword::test::MakeModel( "entity", identifier )
-                                                 ( "code", MIL_Report::eReport_CiviliansEncountered ) );
+                                                 ( "code", static_cast< int >( MIL_Report::eReport_CiviliansEncountered ) ) );
     PostCommand( "perception", core::MakeModel( "identifier", identifier ) );
     ExecuteCommands();
 }
